requirements.txt
africastalking==1.2.3
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==20.2.0
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
contextlib2==0.6.0.post1
coverage==5.3
Django==3.1.1
django-environ==0.4.5
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
drf-oidc-auth==0.10.0
future==0.18.2
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.0.1
isort==5.5.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.5.0
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pycryptodomex==3.9.8
pyjwkest==1.4.2
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pytest==6.0.2
pytz==2020.1
regex==2020.7.14
requests==2.24.0
schema==0.7.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wrapt==1.12.1
zipp==3.2.0
